+++ GMC_Run/algo1.h
@@ -37,8 +37,8 @@ public:
 	vector<vector<float>> pos_list;
 	vector<vector<vector<vector<int>>>> spin_motif_groups;
 	vector<vector<vector<vector<int>>>> chem_motif_groups;
-	map <string, float> rule_map_chem;
-	map <string, float> rule_map_spin;
+	map <string, double> rule_map_chem;
+	map <string, double> rule_map_spin;
 
 	Algo1(void);
 	Algo1(Session& _session, SimCell& _sim_cell);
@@ -47,12 +47,12 @@ public:
 	void fill_SMG(vector<vector<int>>& neigh_ind_list);
     void print_state(string contcar_name, int temp);
 	bool bc_check(vector<float> check_vect, vector<float>& pos);
-	float eval_lat();
-	float eval_lat_spin();
-	float eval_site_spin(int site);
-	float eval_site_chem(int site);
-	float eval_spin_flip(int site, float old_spin);
-	float eval_atom_flip(int site);
+    double eval_lat();
+    double eval_lat_spin();
+    double eval_site_spin(int site);
+    double eval_site_chem(int site);
+    double eval_spin_flip(int site, float old_spin);
+    double eval_atom_flip(int site);
 };
 
 #endif
